el-social-bodega/backend/requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic[email]
pydantic-settings
python-multipart
python-jose[cryptography]
pandas
openpyxl
reportlab
httpx
